reader.go: skip now operates in a cleaner way like in cpp-proio

+++ tools/proio-ls/main.go
@@ -67,7 +67,7 @@ func main() {
 	if *event >= 0 {
 		singleEvent = true
 		startingEvent = *event
-		if _, err = reader.Skip(*event); err != nil {
+		if _, err = reader.Skip(uint64(*event)); err != nil {
 			log.Fatal(err)
 		}
 	}
